made barycenter_solvers and network_simplex hidden + deprecated ot.lp.cvx
