+++ RELEASES.md
@@ -6,7 +6,7 @@
 - Implement CG solvers for partial FGW (PR #687)
 - Added feature `grad=last_step` for `ot.solvers.solve` (PR #693)
 - Automatic PR labeling and release file update check (PR #704)
-- Reorganize sub-module  `ot/lp/__init__.py` into separate files. (PR #714) (PR #714)
+- Reorganize sub-module `ot/lp/__init__.py` into separate files (PR #714)
 
 #### Closed issues
 - Fixed `ot.mapping` solvers which depended on deprecated `cvxpy` `ECOS` solver (PR #692, Issue #668)
