+++ ot/lp/__init__.py
@@ -9,10 +9,10 @@ Solvers for the original linear program OT problem.
 # License: MIT License
 
 from . import cvx
-from .cvx import barycenter
 from .dmmot import dmmot_monge_1dgrid_loss, dmmot_monge_1dgrid_optimize
-from .network_simplex import emd, emd2
-from .barycenter_solvers import (
+from ._network_simplex import emd, emd2
+from ._barycenter_solvers import (
+    barycenter,
     free_support_barycenter,
     generalized_free_support_barycenter,
 )
